chore: simplifies code by introducing units and path aliases
By using path aliases components can rely on absolute paths which makes import statements better understandable and readable. For that we needed to extend the storybook main configuration to support path replacements (see comments in the file).

This commit also introduces `units` that replaces common units like
* Length: a length value in a certain unit of length
* Size2d: represents the lengths of 2 sides

These are constructed in object oriented matter to also be able to add different units of lengths and sizes if need be and also common representations (for instance color representations: '0,0,0' , rgb(0,0,0), #00000, ...)